templates/auth/requirements.txt
reactpy
jwt
PyJWT
python-dotenv
passlib
argon2_cffi